add react-router and refactor frontend boilerplate to use it
