+++ frontend/src/component/cars.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Cars() {
+  return <h1>Welcome to the world of Geeks!</h1>;
+}
+
+export default Cars;
