+++ frontend/src/component/events.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Events() {
+  return <h1>Welcome to the world of Geeks!</h1>;
+}
+
+export default Events;
